# Conflicts:
#	Dockerfile
#	docker-compose.yml
#	gata-frontend/.eslintrc
#	gata-frontend/app/old-app/components/ExternalUserMenuItem.tsx
#	gata-frontend/app/old-app/components/GataReportFormDialog.tsx
#	gata-frontend/app/old-app/components/NavLinkStyled.tsx
#	gata-frontend/app/old-app/components/ResponsiveAppBar.tsx
#	gata-frontend/app/old-app/components/RichTextEditor/Image.tsx
#	gata-frontend/app/old-app/components/RichTextEditor/RichTextEditor.tsx
#	gata-frontend/app/old-app/components/RichTextEditor/RichTextElement.tsx
#	gata-frontend/app/old-app/components/UserMenu.tsx
#	gata-frontend/app/old-app/pages/ReportInfoPage/ReportInfoPage.tsx
#	gata-frontend/app/old-app/pages/member/MemberInfo/UserSubscribe.tsx
#	gata-frontend/app/old-app/pages/member/components/LinkExternalUserToGataUserSelect.tsx
#	gata-frontend/app/old-app/router.tsx
#	gata-frontend/app/old-app/types/Auth0Role.type.ts
#	gata-frontend/app/old-app/types/Auth0User.type.ts
#	gata-frontend/package-lock.json
#	gata-frontend/package.json
#	gata-frontend/src/components/RichTextEditor/withImages.tsx
#	gata-frontend/vite.config.ts
#	src/main/kotlin/no/gata/web/ControllerAdvisor.kt
#	src/main/kotlin/no/gata/web/service/Auth0RestService.kt

+++ src/main/kotlin/no/gata/web/controller/ResponsibilityRestController.kt
@@ -9,10 +9,17 @@ import no.gata.web.repository.ResponsibilityYearRepository
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
 import org.springframework.security.access.prepost.PreAuthorize
-import org.springframework.web.bind.annotation.*
+import org.springframework.web.bind.annotation.DeleteMapping
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.PutMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.ResponseStatus
+import org.springframework.web.bind.annotation.RestController
 import org.springframework.web.server.ResponseStatusException
-import java.time.Year
-import java.util.*
+import java.util.UUID
 
 @RestController
 @RequestMapping("api/responsibility")
@@ -25,39 +32,38 @@ class ResponsibilityRestController {
 
     @GetMapping
     @PreAuthorize("hasAuthority('member')")
-    fun getResponsibilities(): List<Responsibility> {
-        return responsibilityRepository.findAll();
+    fun getResponsibilities(): List<DtoOutResponsibility> {
+        return responsibilityRepository.findAll().map { DtoOutResponsibility(it) }
     }
 
     @GetMapping("{responsibilityId}")
     @PreAuthorize("hasAuthority('member')")
-    fun getResponsibility(@PathVariable responsibilityId: String): Responsibility {
-        return responsibilityRepository.findById(UUID.fromString(responsibilityId)).orElseThrow { ResponsibilityNotFound(responsibilityId) }
-    }
-
-    @GetMapping("available")
-    @PreAuthorize("hasAuthority('member')")
-    fun getAvailableResponsibilities(@RequestParam("year") yearParam: String?): List<DtoOutResponsibility> {
-        val allResponsibilities = responsibilityRepository.findAll();
-        val year = if (yearParam == null) Year.now() else Year.of(yearParam.toInt());
-        val responsibilityYears = responsibilityYearRepository.findResponsibilityYearsByYearEquals(year);
-        return allResponsibilities
-                .filter { responsibilityYears.find { responsibilityYear -> responsibilityYear.responsibility?.id == it.id } == null }
-                .map { DtoOutResponsibility(it) }
+    fun getResponsibility(
+        @PathVariable responsibilityId: String,
+    ): DtoOutResponsibility {
+        return responsibilityRepository.findById(UUID.fromString(responsibilityId))
+            .orElseThrow { ResponsibilityNotFound(responsibilityId) }.let { DtoOutResponsibility(it) }
     }
 
     @PostMapping
     @PreAuthorize("hasAuthority('admin')")
-    fun postResponsibility(@RequestBody body: DtoInnResponsibility) {
+    fun postResponsibility(
+        @RequestBody body: DtoInnResponsibility,
+    ) {
         responsibilityRepository.save(
-                Responsibility(id = null, name = body.name, description = body.description, responsibilityYears = null)
+            Responsibility(id = null, name = body.name, description = body.description, responsibilityYears = null),
         )
     }
 
     @PutMapping("{responsibilityId}")
     @PreAuthorize("hasAuthority('admin')")
-    fun putResponsibility(@RequestBody body: DtoInnResponsibility, @PathVariable responsibilityId: String) {
-        val responsibility = responsibilityRepository.findById(UUID.fromString(responsibilityId)).orElseThrow { ResponsibilityNotFound(responsibilityId) }
+    fun putResponsibility(
+        @RequestBody body: DtoInnResponsibility,
+        @PathVariable responsibilityId: String,
+    ) {
+        val responsibility =
+            responsibilityRepository.findById(UUID.fromString(responsibilityId))
+                .orElseThrow { ResponsibilityNotFound(responsibilityId) }
         responsibility.description = body.description
         responsibility.name = body.name
         responsibilityRepository.save(responsibility)
@@ -66,9 +72,16 @@ class ResponsibilityRestController {
     @DeleteMapping("{responsibilityId}")
     @PreAuthorize("hasAuthority('admin')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    fun deleteResponsibility(@PathVariable responsibilityId: String) {
-        val responsibility = responsibilityRepository.findById(UUID.fromString(responsibilityId)).orElseThrow { ResponsibilityNotFound(responsibilityId) }
+    fun deleteResponsibility(
+        @PathVariable responsibilityId: String,
+    ) {
+        val responsibility =
+            responsibilityRepository.findById(UUID.fromString(responsibilityId))
+                .orElseThrow { ResponsibilityNotFound(responsibilityId) }
+        val responsibilityYears = responsibilityYearRepository.findResponsibilityYearsByResponsibility(responsibility)
+        if (responsibilityYears.isNotEmpty()) {
+            throw throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Ansvarsposten er i bruk")
+        }
         responsibilityRepository.delete(responsibility)
     }
-
 }
